Fix autocomplete not removing package for pseudocode commands
