+++ src/nl/hannahsten/texifyidea/completion/handlers/LatexNoMathInsertHandler.kt
@@ -24,6 +24,7 @@ import nl.hannahsten.texifyidea.util.magic.TypographyMagic
 class LatexNoMathInsertHandler(val arguments: List<Argument>? = null) : InsertHandler<LookupElement> {
 
     override fun handleInsert(context: InsertionContext, item: LookupElement) {
+        removeWhiteSpaces(context)
         val command = item.`object` as LatexCommand
 
         when (command.command) {
@@ -89,6 +90,20 @@ class LatexNoMathInsertHandler(val arguments: List<Argument>? = null) : InsertHa
 
         TemplateManager.getInstance(context.project)
             .startTemplate(context.editor, parameterTemplate)
+    }
+
+    /**
+     * Remove whitespaces and everything after that that was inserted by the lookup text.
+     */
+    private fun removeWhiteSpaces(context: InsertionContext) {
+        val editor = context.editor
+        val document = editor.document
+        val offset = editor.caretModel.offset
+        // context.startOffset is the offset of the start of the just inserted text.
+        val insertedText = document.text.substring(context.startOffset, offset)
+        val indexFirstSpace = insertedText.indexOfFirst { it == ' ' }
+        if (indexFirstSpace == -1) return
+        document.deleteString(context.startOffset + indexFirstSpace, offset)
     }
 
     /**
